experiment

+++ include/bp_tree/bp_internal_node.hpp
@@ -205,7 +205,7 @@ namespace stool
             void __increment_a_value_of_sum_deque(uint64_t pos, int64_t value){                
                 #if DEBUG
                 if(value < 0){
-                    assert(this->children_value_sum_deque_.at(pos) >=  -value);
+                    assert((int64_t)this->children_value_sum_deque_.at(pos) >=  -value);
                 }
                 #endif
